fix: ClassMapGenerator get too many classes
Part of #12

Temp fix in waiting of official recommendation:
https://github.com/composer/class-map-generator/issues/6
