+++ tests/unit/Fixtures/brick1/src/MyEvent.php
@@ -25,9 +25,10 @@
 
 declare(strict_types=1);
 
-namespace Archict\Core\Fixtures\brick1;
+namespace Archict\Core\Fixtures\brick1\src;
 
-final class SomeClass
+final class MyEvent
 {
-
+    public bool $listened   = false;
+    public int $nb_listened = 0;
 }
